+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Offline mode support added (load everything from local sources)
 - Thread view added
 - (Optional) Remove bookmarked tweets after download using `--danger-remove-bookmarks`
+- Output colors and additional error outputs added
 
 ### Breaking changes
 - NaN